+++ OpenEstate-IS24-REST-hc42/src/main/java/org/openestate/is24/restapi/hc42/HttpComponents42Client.java
@@ -57,6 +57,8 @@ import org.apache.http.util.EntityUtils;
 import org.openestate.is24.restapi.AbstractClient;
 import org.openestate.is24.restapi.utils.RequestMethod;
 import org.openestate.is24.restapi.utils.Response;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * An implementation of {@link AbstractClient}, that makes use of
@@ -69,7 +71,7 @@ import org.openestate.is24.restapi.utils.Response;
  */
 public class HttpComponents42Client extends AbstractClient
 {
-  //private final static Logger LOGGER = LoggerFactory.getLogger( HttpComponents42Client.class );
+  private final static Logger LOGGER = LoggerFactory.getLogger( HttpComponents42Client.class );
   private transient HttpClient httpClient = null;
 
   /**