+++ OpenEstate-IS24-REST/src/main/java/org/openestate/is24/restapi/utils/SslUtils.java
@@ -24,6 +24,8 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSession;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * SSL helper methods.
@@ -33,7 +35,7 @@ import javax.net.ssl.X509TrustManager;
  */
 public final class SslUtils
 {
-  //private final static Logger LOGGER = LoggerFactory.getLogger( SslUtils.class );
+  private final static Logger LOGGER = LoggerFactory.getLogger( SslUtils.class );
 
   private SslUtils()
   {